+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Mobileraker - Changelog
 
-## [2.9.9] - 2026-05-xx
+## [2.9.10] - 2026-05-31
 
 ### Enhancements
 
@@ -8,6 +8,7 @@
   multiple machines at once! This can be done via the file manager on one "Source" machine.
 
   *Note: This feature is available exclusively for "Supporters".*
+- **Backend Rework**: The app's core framework was updated and should now be more stable and faster.
 - **Print Error Reset**: Added a reset button to the Machine Status Card on the dashboard when a print is in an error
   state, making it consistent with the overview page.